+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- tasks around MySQL slow log
+
 ### Fixed
 - set group permissions for /var/www in Chef cookbook (Vagrant)
 - set proper gem homepage